% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afscOM-package.R
\docType{package}
\name{afscOM-package}
\alias{afscOM}
\alias{afscOM-package}
\title{afscOM: General Fisheries Simulation Model}
\description{
A general age-sex structured, spatially explicit fisheries simulation model.
}
\author{
\strong{Maintainer}: Benjamin Williams \email{ben.williams@noaa.gov}

Authors:
\itemize{
  \item Joshua Zahner \email{jazahner@alaska.edu}
  \item Matt Cheng \email{lhcheng@alaska.edu}
  \item Maia Kapur
}

Other contributors:
\itemize{
  \item Dan Goethel [contributor]
  \item Curry Cunningham [contributor]
  \item Chris Lunsford [contributor]
}

}
\keyword{internal}
