% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_control_rules.R
\name{apply_harvest_control_rule}
\alias{apply_harvest_control_rule}
\title{Apply a harvest control rule}
\usage{
apply_harvest_control_rule(model_dimensions, hcr_func, hcr_pars)
}
\arguments{
\item{model_dimensions}{a named list indicating the dimensions of the model}

\item{hcr_func}{the type of control rule}

\item{hcr_pars}{the associated parameters}
}
\description{
Apply a harvest control rule
}
