% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{apportion_recruitment}
\alias{apportion_recruitment}
\title{Apportion Recruitment}
\usage{
apportion_recruitment(rec_timeseries, apportionment, nyears, nregions)
}
\description{
Apportion Recruitment
}
