% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{apportion_recruitment_single}
\alias{apportion_recruitment_single}
\title{Apportion Recruitment Single}
\usage{
apportion_recruitment_single(recruits, apportionment, nregions)
}
\description{
Apportion Recruitment Single
}
