% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assessment}
\alias{assessment}
\title{Sablefish assessment}
\format{
A list containing assessment data:
}
\usage{
assessment
}
\description{
A dataset containing sablefish assessment results
}
\keyword{datasets}
