% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{beverton_holt}
\alias{beverton_holt}
\title{Beverton-Holt Stock Recruit Relationship}
\usage{
beverton_holt(naa, dem_params, h, R0, S0, sigR)
}
\arguments{
\item{naa}{numbers-at-age array (dimensions [1, nages, nsexes, nregions])}

\item{dem_params}{demographic parameters list susbet to a 
single year (dimensions [1, nages, nsexes, nregions, nfleets])}
}
\description{
Classic beverton-holt SRR parameterized using steepness,
unfished recruitment (R0), and unfished spawning biomass
(S0).
}
