% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_points.R
\name{calculate_spr_refpoints}
\alias{calculate_spr_refpoints}
\title{Calculate F/B Reference Points via SPR Methods}
\usage{
calculate_spr_refpoints(nages, dem_params, joint_selret, spr_target, rec)
}
\arguments{
\item{nages}{number of ages in age structure}

\item{dem_params}{list of demographic parameters
subset to a single year}

\item{joint_selret}{list of total selectivity and
retention across all fleets, weighted by relative
fleet-specific fishing mortality (compute using
`calculate_joint_selret`)}

\item{spr_target}{desired SPR proportion}

\item{rec}{recruitment used to calculate BRP}
}
\value{
list of Fref and Bref
}
\description{
Calculate fishing mortality and biological
fisheries reference points via traditional SPR
methods.
}
\examples{
\dontrun{
calculate_spr_refpoints(nages, dem_params, joint_selret, spr_target, rec)
}
}
