% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_utils.R
\name{catch.at.age}
\alias{catch.at.age}
\title{Catch-at-age}
\usage{
catch.at.age(faa, naa, waa, mort)
}
\arguments{
\item{faa}{instantaneous fishing-mortality-at-age}

\item{naa}{numbers-at-age vector}

\item{waa}{weight-at-age vector}

\item{mort}{instantaneous natural-mortality-at-age}
}
\description{
Computes catch-at-age given a vector of
fishing-mortality-at-age.
}
