% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_quantities.R
\name{compute_bio}
\alias{compute_bio}
\title{Compute Total Stock Biomass}
\usage{
compute_bio(naa, dem_params)
}
\arguments{
\item{naa}{numbers-at-age matrix [nyears, nages, nsexes, nregions]}

\item{dem_params}{demagraphic parameters list}
}
\description{
Compute biomass from a matrix of NAA and a set of demographic
parameters. This computes the sum of male and female biomass
across all age classes. For mature female biomass, see
`compute_ssb`.
}
