% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_quantities.R
\name{compute_fleet_catch}
\alias{compute_fleet_catch}
\title{Compute Catch by Fishing Fleet}
\usage{
compute_fleet_catch(caa)
}
\arguments{
\item{caa}{catch-at-age matrix [nyears, nages, nsexes, nregions, nfleets]}
}
\description{
Compute total catch volume (in mt) for each fishing
fleet from a matrix of CAA.
}
