% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_quantities.R
\name{compute_so}
\alias{compute_so}
\title{Compute Spawning Stock Output}
\usage{
compute_so(naa, dem_params, use_fec = TRUE)
}
\arguments{
\item{naa}{numbers-at-age matrix [nyears, nages, nsexes, nregions]}

\item{dem_params}{demagraphic parameters list}
}
\description{
Compute SO (egg production) from a matrix of NAA and a set 
of demographic parameters.
}
