% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_quantities.R
\name{compute_total_catch}
\alias{compute_total_catch}
\title{Compute Total Catch}
\usage{
compute_total_catch(caa)
}
\arguments{
\item{caa}{catch-at-age matrix [nyears, nages, nsexes, nregions, nfleets]}
}
\description{
Compute total catch volume (in mt) from a matrix of CAA.
}
