% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_quantities.R
\name{compute_total_f}
\alias{compute_total_f}
\title{Compute Total Fishing Mortality}
\usage{
compute_total_f(faa)
}
\arguments{
\item{faa}{fushing-mortality-at-age matrix [nyears, nages, nsexes, nregions, nfleets]}
}
\description{
Compute total fishing mortality from a matrix of FAA.
Note that this is an approximation of total fishing mortality,
and assumes that selectivity = 1 for some age class.
}
