% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_control_rules.R
\name{constant_catch_hcr}
\alias{constant_catch_hcr}
\title{Constant Catch Harvest Control Rule}
\usage{
constant_catch_hcr(c, ...)
}
\arguments{
\item{c}{constant catch level}
}
\value{
c
}
\description{
Return a constant level of harvest (c) in
catch units.
}
