% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_control_rules.R
\name{constant_f_hcr}
\alias{constant_f_hcr}
\title{Constant Fishing Mortality Harvest Control Rule}
\usage{
constant_f_hcr(f, ...)
}
\arguments{
\item{f}{constant catch level}
}
\value{
f
}
\description{
Return a constant level of fishing mortality (F) in
catch units.
}
