% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{cyclic_recruitment}
\alias{cyclic_recruitment}
\title{Generate log-recruitment from two defined regimes of a specified length}
\usage{
cyclic_recruitment(n, mu_rec, sd_rec, dur, start = 1, seed = NA)
}
\arguments{
\item{n}{number of recruitment events}

\item{mu_rec}{vector of mean log-recruitment levels, one for each regime}

\item{sd_rec}{vector of recruitment variation, one for each regime}

\item{dur}{duration of regimes}

\item{start}{which regime to start in}

\item{seed}{a random seed (optional)}
}
\value{
`n` log-recruitment levels from two different regimes of
length `dur`.
}
\description{
Description
}
