% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_utils.R
\name{discard.F}
\alias{discard.F}
\title{Discarded Fishing Mortality Rate}
\usage{
discard.F(dmr, selex, ret)
}
\arguments{
\item{dmr}{the instanteous discard mortality rate}

\item{selex}{selectivity-at-age}

\item{ret}{retention-at-age}
}
\description{
Calculates the instantaneous discard mortality rate -at-age
}
