% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_utils.R
\name{discard_F}
\alias{discard_F}
\title{Discarded Fishing Mortality Rate}
\usage{
discard_F(dmr, selex, ret)
}
\arguments{
\item{dmr}{the instanteous discard mortality rate}

\item{selex}{selectivity-at-age}

\item{ret}{retention-at-age}
}
\value{
discard-mortality-at-age (selex*(1-ret)*dmr)
}
\description{
Calculates the instantaneous discard-mortality-at-age
}
