% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_years}
\alias{extend_years}
\title{Extend demographic parameter matrices to more years}
\usage{
extend_years(dem_params, dimension, e, new.dimnames = NA)
}
\arguments{
\item{dem_params}{a list of demographic parameter matrices}

\item{dimension}{the dimension along which to extend (should always be 1)}

\item{e}{the new number of years along the first dimension}

\item{new.dimnames}{new set of names for the first dimension}
}
\value{
a new list of demographic parameter matrices spanning `e` years
with the values from the last year in the original matrix continued for
all future years
}
\description{
Description
}
