% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{F_to_mu}
\alias{F_to_mu}
\title{instantaneous rate to harvest rate}
\usage{
F_to_mu(fy)
}
\arguments{
\item{fy, }{an instantaneous fishing mortality rate}
}
\value{
mu, the corresponding harvest rate
}
\description{
Converts between instantaneous fishing mortality
rate F and harvest rate -the proportion of the
population harvest; mu_
}
