% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findF.R
\name{findF_bisection}
\alias{findF_bisection}
\title{Find F given a desired level of catch}
\usage{
findF_bisection(f_guess, naa, waa, mort, selex, ret = NA, dmr = NA, prov_catch)
}
\arguments{
\item{f_guess}{initial guess for F}

\item{naa}{numbers-at-age}

\item{waa}{weight-at-age}

\item{mort}{natural-mortality-at-age}

\item{selex}{selectivity-at-age}

\item{ret}{retention-at-age (optional)}

\item{dmr}{instantaneous discard mortality rate (optional)}

\item{prov_catch}{desired catch level}
}
\value{
the fishing mortality rate that yields the desired catch level
}
\description{
Uses the Baranov catch equation to compute the instantenous
fishing mortality rate (F) that would yield a specific level of
catch using bisection
}
\examples{
\dontrun{
findF_bisection(f_guess=0.1, naa=naa, waa=waa, mort=mort, selex=selex, prov_catch=12)
}

}
