% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findF.R
\name{findF_multifleet}
\alias{findF_multifleet}
\title{Find F for Multiple Fleets
Uses multi-fleet Baranov catch equation to compute the instantaneous
fishing mortality rate (F) that would yield a specific level of
catch_}
\usage{
findF_multifleet(target_catch, naa, waa, mort, selex, f_guess = 0.05)
}
\arguments{
\item{target_catch}{desired catch level}

\item{naa}{numbers-at-age}

\item{waa}{weight-at-age}

\item{mort}{natural-mortality-at-age}

\item{selex}{selectivity-at-age}

\item{f_guess}{initial guess for F}

\item{ret}{retention-at-age (optional)}

\item{dmr}{instantaneous discard mortality rate (optional)}
}
\value{
the fishing mortality rate that yields the desired catch level
}
\description{
Find F for Multiple Fleets
Uses multi-fleet Baranov catch equation to compute the instantaneous
fishing mortality rate (F) that would yield a specific level of
catch_
}
\examples{
\dontrun{
find_F(f_guess=0.1, naa=naa, waa=waa, mort=mort, selex=selex, prov_catch=0.12)
}

}
