% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_output_matrices}
\alias{generate_output_matrices}
\title{Generate output matrices}
\usage{
generate_output_matrices(nyears, nages, nsexes, nregions, nfleets, nsurveys)
}
\arguments{
\item{nyears}{number of years}

\item{nages}{number of ages}

\item{nsexes}{number of sexes}

\item{nregions}{number of regions}

\item{nfleets}{number of fishing fleets}

\item{nsurveys}{number of surveys}
}
\description{
Generate output matrices
}
