% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{get_annual_recruitment}
\alias{get_annual_recruitment}
\title{Get Annual Recruitment}
\usage{
get_annual_recruitment(
  recruitment,
  apportionment,
  apportion_random,
  apportionment_pars,
  nregions
)
}
\description{
Get Annual Recruitment
}
