% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_dimensions.R
\name{get_model_dimensions}
\alias{get_model_dimensions}
\title{Obtain model dimensions from parameter matrix}
\usage{
get_model_dimensions(par_mat)
}
\arguments{
\item{par_mat}{a parameter matrix generated via generate_param_matrix}
}
\value{
a named list indicating the dimensions of the model
}
\description{
Automatically determines nyears, nages, nsexes, nregions, and nfleets
based on the dimensions of the input parameter matrix. This should be
run on the parameter matrix with the largest number of dimensions but
will return the correct dimensions for a smaller matrix.
}
