% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{good_dem_matrices}
\alias{good_dem_matrices}
\title{good_dem_matrices}
\format{
A list of data inputs
}
\usage{
good_dem_matrices
}
\description{
Demographic parameter inputs
}
\keyword{datasets}
