% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kapur2024_sablefish_data}
\alias{kapur2024_sablefish_data}
\title{kapur2024_sablefish_data}
\format{
A list of data inputs
}
\usage{
kapur2024_sablefish_data
}
\description{
Kapur et al. dataset
}
\keyword{datasets}
