% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listN}
\alias{listN}
\title{Set variable names as names of list}
\usage{
listN(...)
}
\arguments{
\item{...}{R variables containung values to be put in a list}
}
\description{
Automatically use variable names as list object names when
a sequence of variables is provided to a list constructor_
}
