% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mu_to_F}
\alias{mu_to_F}
\title{harvest rate to instantaneous}
\usage{
mu_to_F(mu)
}
\arguments{
\item{mu, }{a harvest rate}
}
\value{
F, the corresponding instantaneous fishing mortality rate
}
\description{
Converts between harvest rate - the proportion
of the population harvest; mu and instantaneous
fishing mortality rate F_
}
