% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bio}
\alias{plot_bio}
\title{Plot Total biomass}
\usage{
plot_bio(bio, compare_ts = NULL)
}
\arguments{
\item{bio}{annual total biomass}

\item{compare_ts}{i don't know what this does - default:NULL}
}
\description{
Plot Total biomass
}
