% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot Catch}
\usage{
plot_catch(catch, compare_ts = NULL)
}
\arguments{
\item{catch}{annual catch}

\item{compare_ts}{i don't know what this does - default:NULL}
}
\description{
Plot Catch
}
