% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_demographic_parameters}
\alias{plot_demographic_parameters}
\title{Plot demographic parameters}
\usage{
plot_demographic_parameters(
  dem_params,
  params = NA,
  show_plots = TRUE,
  out_dir = NA,
  ...
)
}
\arguments{
\item{dem_params}{list of demographic parameter matrices}

\item{params}{vector of names of parameters to plot (not implemented)}

\item{out_dir}{optional directory to save plots to}

\item{...}{additional parameters to be passed to `ggsave`}
}
\description{
Generate standard plot for all demographic
parameter matrices. Output plots will include:
- weight-at-age
- maturity
- natural mortality
- selectivity
- retention
- discard mortality
Plots will show values of demographic quantities across
time, age, sex, region, and fleet as appropriate.
}
