% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_f}
\alias{plot_f}
\title{Plot Fishing mortality biomass}
\usage{
plot_f(f, compare_ts = NULL)
}
\arguments{
\item{f}{fishing mortality}

\item{compare_ts}{i don't know what this does - default:NULL}
}
\description{
Plot Fishing mortality biomass
}
