% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mat}
\alias{plot_mat}
\title{Plot Maturity-at-Age}
\usage{
plot_mat(mat)
}
\arguments{
\item{mat}{four-dimensional maturity-at-age matrix}
}
\description{
Generates plot of maturity-at-age across time, age, sex,
and region as appropriate. Only the female maturity ogive
is plotted.
}
