% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mort}
\alias{plot_mort}
\title{Plot Mortality}
\usage{
plot_mort(mort, is_dmr = FALSE)
}
\arguments{
\item{waa}{four-dimensional mortality-at-age matrix}
}
\description{
Generates plot of mortality-at-age across time, age, sex,
and region as appropriate. This function can be used for
both natural and discard mortality
}
