% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_selret}
\alias{plot_selret}
\title{Plot Selectivity-at-Age and Retention-at-Age}
\usage{
plot_selret(selret, is_selectivity = TRUE)
}
\arguments{
\item{selret}{five-dimensional selectivity-at-age or
retention-at-age matrix}

\item{is_selectivity}{default: TRUE, otherwise retention is output}
}
\description{
Generates plot of selectivity-at-age or retention-at-age
across time, age, sex, region, and fleet as appropriate.
}
