% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ssb}
\alias{plot_ssb}
\title{Plot Spawning stock biomass}
\usage{
plot_ssb(ssb, compare_ts = NULL)
}
\arguments{
\item{ssb}{annual spawning biomass}

\item{compare_ts}{i don't know what this does - default:NULL}
}
\description{
Plot Spawning stock biomass
}
