% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_waa}
\alias{plot_waa}
\title{Plot Weight-at-Age}
\usage{
plot_waa(waa)
}
\arguments{
\item{waa}{four-dimensional weight-at-age matrix}
}
\description{
Generates plot of weight-at-age across time, age, sex,
and region as appropriate.
}
