% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project population forward by 1 year}
\usage{
project(removals, fleet.props, dem_params, prev_naa, recruitment, options = NA)
}
\arguments{
\item{removals}{population removals (either in catch or F units)}

\item{fleet.props}{the proportion of the TAC allocated to each fleet}

\item{dem_params}{list of demographic parameter matrices}

\item{prev_naa}{the NAA in the previous timestep}

\item{recruitment}{total recruiment to occurin this year}

\item{option}{additional model options}
}
\value{
list of derived quantities included landed catch-at-age,
discarded catch-at-age, total catch-at-age, F-at-age, and numbers-at-age.
}
\description{
Given a TAC, the previous population age structure, demographic information
(e.g. natural mortality, selectivity, retention, etc.), and recruitment,
determine the numbers-at-age present in the population for the following year.
}
