% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_multi.R
\name{project_multi}
\alias{project_multi}
\title{Project Population Forward Multiple Years}
\usage{
project_multi(
  init_naa,
  removals_timeseries,
  recruitment,
  dem_params,
  nyears,
  model_options
)
}
\arguments{
\item{init_naa}{numbers-at-age matrix in starting year ([1, nages, nsexes, nregions])}

\item{removals_timeseries}{vector of removals (either catch of F) of length nyears}

\item{recruitment}{vector of recruitmene of length nyears}

\item{dem_params}{list of demographic parameter matrices}

\item{nyears}{number of projection yeas}

\item{model_option}{list of additional model options}
}
\description{
A wrapper function around the base `project` function that handles
projecting forward multiple years given a removals timeseries and a 
recruitment timeseries.
}
