% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_sablefish_dem_params}
\alias{raw_sablefish_dem_params}
\title{raw_sablefish_dem_params}
\format{
A list of demographic parameters
}
\usage{
raw_sablefish_dem_params
}
\description{
raw_sablefish_dem_params
}
\keyword{datasets}
