% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{resample_recruitment}
\alias{resample_recruitment}
\title{Resample from historical recruitment timeseries}
\usage{
resample_recruitment(n, hist_rec, seed = NA)
}
\arguments{
\item{n}{number of samples to draw}

\item{hist_rec}{historical recruitment vector}

\item{seed}{a random seed (optional)}
}
\value{
`n` random samples from the historical
recruitment timeseries.
}
\description{
Draw `n` random samples, with replacement, from the
historical recruitment timeseries.
}
\examples{
set.seed(1120)
n <- 100
hist.rec <- c(100, 200, 50, 100, 120)
resample.recruitment(n, hist.rec)

}
