% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_utils.R
\name{retained.F}
\alias{retained.F}
\title{Retained Fishing Mortality Rate}
\usage{
retained.F(fy, selex, ret)
}
\arguments{
\item{fy}{the instantaneous fishing mortality rate (F)}

\item{selex}{selectivity-at-age}

\item{ret}{retention-at-age}
}
\description{
Calculates the instantaneous fishing mortality rate -at-age
}
