% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_utils.R
\name{retained_F}
\alias{retained_F}
\title{Retained Fishing Mortality Rate}
\usage{
retained_F(fy, selex, ret)
}
\arguments{
\item{fy}{the instantaneous fishing mortality rate (F)}

\item{selex}{selectivity-at-age}

\item{ret}{retention-at-age}
}
\value{
fishing-mortality-at-age (fy*selex*ret)
}
\description{
Calculates the instantaneous fishing-mortality-at-age
}
