% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sablefish_dem_params}
\alias{sablefish_dem_params}
\title{Demographic parameter matrices for Alaska Sablefish}
\format{
A list containing demographic matrices:
}
\usage{
sablefish_dem_params
}
\description{
A dataset containing fully formed demographic
parmeter matrices for Alaska sablefish, based on
the 2023 sablefish assessment.
}
\keyword{datasets}
