% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_model_params.R
\name{set_model_params}
\alias{set_model_params}
\title{Setup model parameters}
\usage{
set_model_params(nyears, nages, nsexes = 1, nregions = 1, nfleets = 1)
}
\arguments{
\item{nyears}{years of input data}

\item{nages}{number of ages}

\item{nsexes}{number of sexes}

\item{nregions}{number of regions}

\item{nfleets}{number of fishing fleets}
}
\description{
Setup model parameters
}
\examples{
\dontrun{
set_model_params(nyears, nages, nsexes=1, nregions=1, nfleets=1)
}
}
