% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_model_options}
\alias{setup_model_options}
\title{Set Default Values for Model Options}
\usage{
setup_model_options(model_dimensions)
}
\arguments{
\item{model_dimensions}{model dimensions list object}
}
\description{
Set up a fully formed model_options list object with all
required elements set to sensible default values.
}
\examples{
\dontrun{
     dimensions = list(nyears=60, nages=50, nsexes=2, nregions=1, nfleets=1)
     model_options = setup_model_options(dimensions)
}

}
