% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_om}
\alias{simple_om}
\title{OM Components for a Simple Simulation}
\format{
A list of OM components
}
\usage{
simple_om
}
\description{
Demographic parameters, initial population structure,
recruitment timeseries, catch timeseries, and model
options for running a simple 25-year age-sex-structured
population simulation with one spatial region and one
fishing fleet. The observation submodel is turned off
in this example
}
\keyword{datasets}
