% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_utils.R
\name{simulate_caa}
\alias{simulate_caa}
\title{Simulate catch-at-age composition data}
\usage{
simulate_caa(naa, faa, zaa, aggregate_sex = FALSE)
}
\arguments{
\item{naa}{numbers-at-age vector}

\item{faa}{fishing-mortality-at-age vector}

\item{zaa}{total-mortality-at-age vector}

\item{aggregate_sex}{whether to return sex-aggregated comps or not (default FALSE)}
}
\description{
Generate an age composition vector based on
fishing induced mortality.
}
