% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_utils.R
\name{simulate_multinomial_obs}
\alias{simulate_multinomial_obs}
\title{Simulate observations from a multinomial distribution
#'
A wrapper function around `rmultinom` that generates a
single vector of observations from a multinomial distribution
where each of the `K` multinomial classes has probability
`pred`. This function will automatically handle generating
multinomial draws across multiple sexes. Aging error can be
optionally applied.}
\usage{
simulate_multinomial_obs(
  pred,
  samp_size,
  aggregate_sex = FALSE,
  as_integers = FALSE,
  age_err = NA
)
}
\arguments{
\item{pred}{the probability associated with each class}

\item{samp_size}{multinomial sample size}

\item{age_err}{whether to automatically apply ageing error}
}
\description{
Simulate observations from a multinomial distribution
#'
A wrapper function around `rmultinom` that generates a
single vector of observations from a multinomial distribution
where each of the `K` multinomial classes has probability
`pred`. This function will automatically handle generating
multinomial draws across multiple sexes. Aging error can be
optionally applied.
}
\examples{
\dontrun{
simulate_multinomial_obs(c(0.25, 0.50, 0.25), 100)
}

}
