% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{simulate_population}
\alias{simulate_population}
\title{Project an Age Structure Population Forward by 1 year
#'
Description}
\usage{
simulate_population(prev_naa, faa, recruitment, dem_params, options)
}
\arguments{
\item{prev_naa}{numbers-at-age vector for previous year}

\item{faa}{fishing mortality-at-age vector (dim [1, nages, nsexes, 1, nfleets])}

\item{recruitment}{projected recruitment for the folliwing year}

\item{dem_params}{demographic parameter matrices subsetted to 1 year and 1 region}

\item{options}{model options list}
}
\description{
Project an Age Structure Population Forward by 1 year
#'
Description
}
