% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_utils.R
\name{simulate_rpn}
\alias{simulate_rpn}
\title{Simulate Relative Population Numbers Index}
\usage{
simulate_rpn(q, naa, sel, zaa)
}
\arguments{
\item{q}{catchability coefficient}

\item{naa}{numbers-at-age vector}

\item{sel}{selectivity-at-age vector}

\item{zaa}{total mortality-at-age vector}
}
\description{
Generate a relative population number index based on
a survey occurring halfway through the year.
}
