% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_utils.R
\name{simulate_rpw}
\alias{simulate_rpw}
\title{Simulate Relative Population Weight Index}
\usage{
simulate_rpw(q, naa, waa, sel, zaa)
}
\arguments{
\item{q}{catchability coefficient}

\item{naa}{numbers-at-age vector}

\item{waa}{weight-at-age vector}

\item{sel}{selectivity-at-age vector}

\item{zaa}{total mortality-at-age vector}
}
\description{
Generate a relative population weight index based on
a survey occurring halfway through the year.
}
