% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_sablefish_assessment_data}
\alias{spatial_sablefish_assessment_data}
\title{Spatial Sablefish Assessment}
\format{
A list of data inputs and model outputs
}
\usage{
spatial_sablefish_assessment_data
}
\description{
Demographic parameter inputs and estimated
derived quantities (selectivity, spawning biomass timeseries,
fishing mortality, etc.) from a run of the SpatialSablefishAssessment
model (https://craig44.github.io/SpatialSablefishAssessment/index.html).
Model developed by Craig Marsh. Data are considered preliminary.
}
\keyword{datasets}
