% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_dem_params}
\alias{subset_dem_params}
\title{Subset demographic parameter matrices by first dimensions}
\usage{
subset_dem_params(dem_params, r, d = 1, drop = TRUE)
}
\arguments{
\item{dem_params}{a list of arrays with more than 2 dimensions}

\item{r}{the index along the first dimension to access}
}
\value{
a list with the same elements as dem_params but containing
a single row from each list element
}
\description{
Access an arbitrary row across all demographic parameter
matrices of indeterminate dimension
}
