% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_matrix}
\alias{subset_matrix}
\title{Subset matrix of undetermined dimensions}
\usage{
subset_matrix(mat, r, d = 1, drop = TRUE)
}
\arguments{
\item{mat}{the input matrix to subset}

\item{r}{the index along the dimensions of interest to subset by}

\item{d}{the dimension to subset by}

\item{drop}{whether to drop the subsetted dimension}
}
\value{
a subsetted matrix or array
}
\description{
Return a subset or a multidimensional matrix or array of undetermined
size along a specific dimension. Specifically designed to return a
single index along the specified dimension (e.g. one year or one region).
}
