% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_control_rules.R
\name{threshold_f_hcr}
\alias{threshold_f_hcr}
\title{Threshold Fishing Mortality Harvest Control Rule}
\usage{
threshold_f_hcr(fmax, fmin, urp, lrp, naa, dem_params)
}
\arguments{
\item{fmax}{maximum F to apply}

\item{fmin}{minimum F to apply}

\item{urp}{SSB level above which fmax is applied}

\item{lrp}{SSB level below which fmin is applied}

\item{naa}{numbers-at-age vector (used to internally compute SSB)}

\item{dem_params}{demographic parameters (used to internally compute SSB)}
}
\value{
F as defined by SSB and the provided reference points
}
\description{
Return an allowable fishing mortality rate based on
a threshold (hockey-stick) harvest control rule, using 
spawning stock biomass (SSB) as the input.
}
