% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_output_matrices}
\alias{update_output_matrices}
\title{update matrices}
\usage{
update_output_matrices(output_matrices, y, input, update_obs = FALSE)
}
\arguments{
\item{output_matrices}{previous OM output}

\item{y}{year}

\item{input}{inputs (land_caa_tmp, caa_tmp, naa_tmp, etc.)}

\item{update_obs}{switch to also update the observation data, default: FALSE}
}
\description{
update matrices
}
\examples{
\dontrun{
update_output_matrices(output_matrices, y, input, update_obs=FALSE)
}
}
