% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baranov.R
\name{baranov}
\alias{baranov}
\title{Baranov Catch Equation}
\usage{
baranov(fy, naa, waa, mort, selex, ret = NA, dmr = NA)
}
\arguments{
\item{fy}{an instantenous fishing mortality rate to apply to the population}

\item{naa}{numbers-at-age in the population}

\item{waa}{weight-at-age vector}

\item{mort}{instanteous natural mortality (M) -at-age vector}

\item{selex}{selectivity-at-age vector}

\item{ret}{retention-at-age vector (optional)}

\item{dmr}{instanteous discard mortality rate -at-age vector (optional)

If `ret` or `dmr` parameters are not provided, retention is assumed to be 1.0
for all ages.}
}
\description{
An implementation of the Baranov catch equations allowing for
fishery retenetion and discarding practices.
}
