% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_quantities.R
\name{compute_ssb}
\alias{compute_ssb}
\title{Compute Spawning Stock Biomass}
\usage{
compute_ssb(naa, dem_params)
}
\arguments{
\item{naa}{numbers-at-age matrix [nyears, nages, nsexes, nregions]}

\item{dem_params}{demagraphic parameters list}
}
\description{
Compute SSB from a matrix of NAA and a set of demographic
parameters.
}
