% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_utils.R
\name{simulate_ac}
\alias{simulate_ac}
\title{Simulate survey age composition data}
\usage{
simulate_ac(naa, sel, aggregate_sex = FALSE)
}
\arguments{
\item{naa}{numbers-at-age vector}

\item{sel}{selectivity-at-age vector}

\item{aggregate_sex}{whether to return sex-aggregated comps or not (default FALSE}
}
\description{
Generate an age composition vector based on
a survey occurring halfway through the year,
and for which ageing error may exist.
}
