% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_catch.R
\name{simulate_catch}
\alias{simulate_catch}
\title{Simulate catches in one region during on year}
\usage{
simulate_catch(removals, dem_params, naa, options)
}
\arguments{
\item{removals}{desired removals either in catch or F units}

\item{dem_params}{demographic parameter matrices for population}

\item{naa}{the current numbers-at-age in the population}

\item{options}{list of model options}
}
\description{
Description
}
\examples{
/dontrun{
simulate_catch(removals, dem_params, naa, options)
}
}
