% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_catch.R
\name{simulate_catch}
\alias{simulate_catch}
\title{Simulate catches in one region during on year}
\usage{
simulate_catch(TAC, fleet.props, dem_params, naa, options)
}
\arguments{
\item{TAC}{the total catch to be removed across all fleets}

\item{fleet.props}{the proportion of the TAC that should be taken by each fleet}

\item{dem_params}{demographic parameter matrices for population}

\item{naa}{the current numbers-at-age in the popultion}

\item{option}{list of model options}
}
\description{
Description
}
