% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_observations.R
\name{simulate_observations}
\alias{simulate_observations}
\title{Simulate Observations from Fisheries and Surveys}
\usage{
simulate_observations(naa, waa, selex, faa, zaa, obs_pars, age_error = NA)
}
\arguments{
\item{$$}{}
}
\description{
Description
}
